pyproject.toml
[project]
name = "bot-reload-aiogram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram>=3.18.0",
    "python-dotenv>=1.0.1",
]
